refactor: move auth factories into cvat_sdk.core.auth, re-export from CLI
