+++ cvat-cli/src/cvat_cli/_internal/common.py
@@ -8,23 +8,16 @@ import getpass
 import importlib
 import importlib.util
 import logging
-import os
 import sys
 import textwrap
-from collections.abc import Callable
 from http.client import HTTPConnection
 from pathlib import Path
 from typing import Any
 
 import attrs
 import cvat_sdk.auto_annotation as cvataa
-from cvat_sdk.core.client import (
-    AccessTokenCredentials,
-    Client,
-    Config,
-    Credentials,
-    PasswordCredentials,
-)
+from cvat_sdk.core.auth import CVAT_ACCESS_TOKEN_ENV_VAR, default_auth_factory, get_auth_factory
+from cvat_sdk.core.client import Client, Config
 
 from ..version import VERSION
 from .parsers import BuildDictAction, parse_function_parameter
@@ -33,42 +26,6 @@ from .utils import popattr
 
 class CriticalError(Exception):
     pass
-
-
-CVAT_ACCESS_TOKEN_ENV_VAR = "CVAT_ACCESS_TOKEN"  # nosec - a variable name declaration
-
-
-def default_auth_factory() -> Callable[[str], Credentials]:
-    """
-    Try to read the CVAT_ACCESS_TOKEN environment variable for a Personal Access Token (PAT).
-    If there is no value, try using the current user and asking for the password.
-    """
-
-    token = os.getenv(CVAT_ACCESS_TOKEN_ENV_VAR)
-    if token is not None:
-        return lambda _: AccessTokenCredentials(token)
-
-    return get_auth_factory(getpass.getuser())
-
-
-def get_auth_factory(s: str) -> Callable[[str], Credentials]:
-    """
-    Parse a USER[:PASS] string and return a callable that takes the server URL
-    and returns auth credentials for that URL.
-    The callable will prompt the user for the password if none was initially supplied in the
-    input string and in the PASS env variable.
-    """
-
-    user, _, password = s.partition(":")
-    if not password:
-        password = os.environ.get("PASS")
-
-    if password:
-        return lambda _: PasswordCredentials(user, password)
-    else:
-        return lambda url: PasswordCredentials(
-            user, getpass.getpass(f"Password for {user} at {url}: ")
-        )
 
 
 def configure_common_arguments(parser: argparse.ArgumentParser) -> None:
